+++ driver/src/main/scala/org/mongodb/scala/ListCollectionsObservable.scala
@@ -58,5 +58,17 @@ case class ListCollectionsObservable[TResult](wrapped: ListCollectionsIterable[T
     this
   }
 
+  /**
+   * Sets the number of documents to return per batch.
+   *
+   * @param batchSize the batch size
+   * @return this
+   * @since 2.7
+   */
+  def batchSize(batchSize: Int): ListCollectionsObservable[TResult] = {
+    wrapped.batchSize(batchSize)
+    this
+  }
+
   override def subscribe(observer: Observer[_ >: TResult]): Unit = observe(wrapped).subscribe(observer)
 }