+++ driver/src/main/scala/org/mongodb/scala/ListDatabasesObservable.scala
@@ -71,5 +71,17 @@ case class ListDatabasesObservable[TResult](wrapped: ListDatabasesIterable[TResu
     this
   }
 
+  /**
+   * Sets the number of documents to return per batch.
+   *
+   * @param batchSize the batch size
+   * @return this
+   * @since 2.7
+   */
+  def batchSize(batchSize: Int): ListDatabasesObservable[TResult] = {
+    wrapped.batchSize(batchSize)
+    this
+  }
+
   override def subscribe(observer: Observer[_ >: TResult]): Unit = observe(wrapped).subscribe(observer)
 }