+++ driver/src/test/scala/org/mongodb/scala/ListCollectionsObservableSpec.scala
@@ -43,6 +43,7 @@ class ListCollectionsObservableSpec extends FlatSpec with Matchers with MockFact
 
     val filter = Document("a" -> 1)
     val duration = Duration(1, TimeUnit.SECONDS)
+    val batchSize = 10
     val observer = new Observer[Document]() {
       override def onError(throwable: Throwable): Unit = {}
       override def onSubscribe(subscription: Subscription): Unit = subscription.request(Long.MaxValue)
@@ -58,6 +59,12 @@ class ListCollectionsObservableSpec extends FlatSpec with Matchers with MockFact
 
     observable.filter(filter)
     observable.maxTime(duration)
+    observable.subscribe(observer)
+
+    wrapper.expects('batchSize)(batchSize).once()
+    wrapper.expects('getBatchSize)().once()
+
+    observable.batchSize(batchSize)
     observable.subscribe(observer)
   }
 }